split tests to separate files
